pdf-extraction-service/requirements.txt
Flask
PyMuPDF
pytesseract
Pillow